backend/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework>=3.0.0
django-cors-headers>=3.0.0
requests>=2.0.0
